[Refactor] Decompose Box transport and command polling (#1290)
